NIO - refactor compilerwork layer
